requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==3.2.0
click==8.1.8
dnspython==2.1.0
email-validator==1.1.3
fastapi==0.68.1
h11==0.16.0
idna==3.10
motor==2.5.1
passlib[bcrypt]==1.7.4
pydantic==1.8.2
pydantic_core==2.33.1
pymongo==3.12.0
python-dotenv==0.19.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
utils==1.0.2
uvicorn==0.15.0
python-multipart==0.0.5
python-jose[cryptography]==3.3.0
python-dateutil==2.8.2
pytz==2021.1
APScheduler==3.8.1
aiohttp==3.8.1
aiodns==3.0.0
cchardet==2.1.7
ujson==4.0.2
Jinja2==3.0.1
requests==2.26.0
google-generativeai==0.3.2
PyJWT==2.4.0
